pyproject.toml
[tool.poetry]
name = "fastapi-project-template"
version = "0.1.0"
description = ""
authors = ["Donald Trieu <donaldtrieu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.28"
alembic = "^1.13.1"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"
psycopg = {extras = ["binary"], version = "^3.1.18"}


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.4"
isort = "^5.13.2"
flake8 = "^7.0.0"
black = "^24.2.0"
pytest = "^8.0.2"
pytest-cov = "^4.1.0"
pytest-forked = "^1.6.0"
pytest-xdist = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''